pyproject.toml
[tool.poetry]
name = "github-pr-contribution-util"
version = "0.1.0"
description = ""
authors = ["MattNotarangelo <matt.notarangelo1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
